% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_fun.R
\name{beta}
\alias{beta}
\title{Calculate slope/beta of an OLS regression using matrix algebra.
This will be the Slope Index of Inequality (SII).}
\usage{
beta(a = 1, x, Y)
}
\arguments{
\item{a}{alpha, default \code{a=1}}

\item{x}{population distribution}

\item{Y}{standardized rates for calculation}
}
\value{
Slope of an OLS regression.
}
\description{
Calculate slope/beta of an OLS regression using matrix algebra.
This will be the Slope Index of Inequality (SII).
}
