% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_multinom_fun.R
\name{CI_multinom}
\alias{CI_multinom}
\title{Function for CI simulation using the multinomial distribution, internal function.}
\usage{
CI_multinom(df, N, n_g)
}
\arguments{
\item{df}{data frame}

\item{N}{Number of simulations}

\item{n_g}{Number of deprivation categories/groups}
}
\value{
Data frame of simulated data
}
\description{
Function for CI simulation using the multinomial distribution, internal function.
}
