% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_long.R
\docType{data}
\name{dep_data_long}
\alias{dep_data_long}
\title{Scottish health and deprivation data - long format}
\format{A data frame with 54648 rows and 24 variables:
\describe{
  \item{PS_code}{postcode sector code}
  \item{PS_name}{postcode sector name}
  \item{health_board}{health board name}
  \item{council_area}{Council area name}
  \item{ur2fold}{Two-fold urban rural classification (created based on output area level data)}
  \item{total_pop}{Total population of the postcode sector}
  \item{pcnt_overcrowding}{Percent of people living in overcrowded households}
  \item{pcnt_male_unemp}{Percent of unemployed men}
  \item{pcnt_no_car}{Percent of people in households with no access to a car/van}
  \item{pcnt_low_class}{Percent of people in low social class}
  \item{carstairs}{The 2011 Carstairs score}
  \item{pcnt_limited_lot}{Percent of people whose day-to-day activities are limited a lot}
  \item{pcnt_no_llti}{Percent of people whose day-to-day activities are not limited}
  \item{pcnt_bad_health}{Percent of people in bad health}
  \item{pcnt_good_health}{Percent of people in good health}
  \item{quintile}{Quintiles of the Carstairs score; 1 - least deprived, 5 - most deprived}
  \item{decile}{Deciles of the Carstairs score; 1 - least deprived, 10 - most deprived}
  \item{quintile_2}{Quintiles of the Carstairs score, factor variable coded in text}
  \item{ethnicity}{Ethnic group, either "all" for all people, "Scot" for Scottish and "asian" for Asians}
  \item{age}{age group, 1 - "0-4", 2 - "5-9", up to 18 - "85 and above"}
  \item{bad}{number of people in bad general health}
  \item{fg}{number of people in fairly good health}
  \item{good}{number of people in good health}
  \item{pop}{number of people in age and ethnic group in postcode sector}
  }}
\source{
For Carstairs deprivation data look at
 \url{https://www.gla.ac.uk/researchinstitutes/healthwellbeing/research/mrccsosocialandpublichealthsciencesunit/programmes/inequalities/carstairsscores/}

For health data see the 2011 census commissioned tables \url{http://www.scotlandscensus.gov.uk/ods-web/data-warehouse.html#additionaltab}
}
\usage{
dep_data_long
}
\description{
A dataset on deprivation and self-rated health outcomes for postcode sectors from 2011 Scottish Census in long format
}
\keyword{datasets}
