% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmid_fun.R
\name{popmid}
\alias{popmid}
\title{Calculates midpoints for population distribution.}
\usage{
popmid(x)
}
\arguments{
\item{x}{vector of population distribution that sums up to 1}
}
\value{
A vector of population midpoints
}
\description{
\code{popmid} calculates the (cumulative) midpoints for population percentiles that can then be used as the
  predictor in the OLS model for calculating the slope or the relative index of inequality (SII).
  The function is mainly for internal use within the \code{rii} function.
}
