% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_fun.R
\name{smr}
\alias{smr}
\title{Standardized mortality ratios and CIs}
\usage{
smr(data, health, population, age, compare, sets, age_group = NULL, CI = 95)
}
\arguments{
\item{data}{Name of dataset.}

\item{health}{Health outcome of interest.}

\item{population}{Population counts. Should correspond to data provided in \code{health}.}

\item{age}{Variable that defines 5-year age groups, should be ordered and numeric, such as 1 through 18 or
0 through 18 if the first age group is under 1 years of age.}

\item{compare}{Categorical variable that splits the data into groups that are to be compared, such as ethnicity.}

\item{sets}{groups that are to be compared, values taken by \code{compare}. The group listed first will be the
reference category. Must take at least two values.}

\item{age_group}{The age groups the standardized rates should be calculated for. By default the function calculates
results for the following age groups: 0-14, 15-29, 30-44, 45-59, 60-74, 75+, 0-64 and all ages.
Results for ages 0-64 and all ages will always be provided.
User supplied age groups should be provided using the standard population groups as cut-offs, e.g.
\code{age_group=c("20-29", "25-39")} and not "19-30" or "21-41". Open ended age groups can be supplied by giving
a single age, e.g. "45" means 45 and above. Overlaping age groups, e.g. \code{c("20-29", "25-34")}, are not supported.}

\item{CI}{Confidence intervals, 95 by default but can be set to any number between 0 and 100.
For calculation method see user guide.}
}
\value{
A data frame of standardized mortality ratios (SMR) and CIs
}
\description{
The function \code{smr} calculates standardized mortality ratios (SMR) and confidence intervals for SMR.
}
\examples{

# Asian population compared to Scottish (reference)
smr(dep_data_long, bad, pop, age, ethnicity, sets = c("Scot", "asian"))

# Asian, White British and Irish population compared to Scottish (reference)
#smr(d_pcs, bad, pop, age, ethnicity, sets = c("Scot", "asian", "WB", "Irish"))

}
