% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rate_fun.R
\name{st_rate}
\alias{st_rate}
\title{Calculates standardized mortality rates and confidence intervals}
\usage{
st_rate(data, health, population, ses, age, groups = NULL, age_group = NULL,
  st_pop = "esp2013_18ag", CI = 95, thousand = 1000)
}
\arguments{
\item{data}{Name of dataset.}

\item{health}{Health outcome of interest.}

\item{population}{Population counts. Should correspond to data provided in \code{health}.}

\item{ses}{Categorical deprivation measure used for splitting the data.}

\item{age}{Variable that defines 5-year age groups, should be ordered and numeric, such as 1 through 18 or
0 through 18 if the first age group is under 1 years of age.}

\item{groups}{Conditions such as sex or ethnicity that together define the sub-population for which the
RII/SII is calculated for. See examples. Default is \code{NULL}, no sub-population is selected.}

\item{age_group}{The age groups the standardized rates should be calculated for. By default the function calculates
results for the following age groups: 0-14, 15-29, 30-44, 45-59, 60-74, 75+, 0-64 and all ages.
User supplied age groups should be provided using the standard population groups as cut-offs, e.g.
\code{age_group=c("20-29", "25-39")} and not "19-30" or "21-41". Open ended age groups can be supplied by giving
a single age, e.g. "45" means 45 and above. Overlaping age groups, e.g. \code{c("20-29", "25-34")}, are not supported.
Results for ages 0-64 and all ages will always be provided.}

\item{st_pop}{the standard population weights used for calculating rates, default 2013 ESP for 18 age groups with 0-4 as
the first age group. See \code{\link{st_pop}} for other predefined options. Can be user supplied (must add up to 1).}

\item{CI}{Confidence intervals, 95 by default but can be set to any number between 0 to 100.}

\item{thousand}{Should rates be calculated per 1000 or 100 000, default 1000}
}
\value{
A data frame giving standardized rates by age and deprivation together with CI.
}
\description{
\code{st_rate} calculates standardized mortality rates and the 95\% or 99\% confidence intervals.
  Standardization uses the 2013 European Standard population.
}
\examples{
d <- dep_data_long

#Standardized rates for all people
st_rate(d, bad, pop, quintile, age, ethnicity=="all")

#Standardized rates for Scottish, with 99\% CI
st_rate(d, bad, pop, quintile, age, ethnicity=="Scot", CI=99)
}
