% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grouping_fun.R
\name{age_grouping}
\alias{age_grouping}
\title{Age grouping function}
\usage{
age_grouping(age_group, st_pop, min_age, n_age)
}
\arguments{
\item{age_group}{character string of age groups for which the results are to be provided.}

\item{st_pop}{standard population weights to use.}

\item{min_age}{minimum age group (to check if it is 0 or 0-4).}

\item{n_age}{number of age groups.}
}
\value{
A small data frame of weights for user selected age groups.
}
\description{
Age grouping function creates standard population weights for user selected age groups.
Allows for flexible age groups selection in smr, rii and st_rate functions. Internal use only.
}
\keyword{internal}
