% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_fun.R
\name{beta}
\alias{beta}
\title{Slope of an OLS regression}
\usage{
beta(x, y, w = NULL)
}
\arguments{
\item{x}{Population distribution.}

\item{y}{Standardized rates for calculation.}

\item{w}{Weight for weighted least squares regression, default \code{NULL}.}
}
\value{
Slope of an OLS regression.
}
\description{
\code{beta} will calculate the slope of an OLS regression using matrix algebra.
This will be the Slope Index of Inequality (SII). Weighted least squares slope is calculated
when weights are supplied.
}
\keyword{internal}
