% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score_fun.R
\name{zscore}
\alias{zscore}
\title{Calculates z-scores}
\usage{
zscore(population, variable)
}
\arguments{
\item{population}{distribution (in numbers)}

\item{variable}{the deprivation measure to calculate the z-scores for}
}
\value{
A list including the following:
 \item{z.score}{standardized score using weighted mean and sd}
 \item{w.mean}{weighted mean}
 \item{w.sd}{weighted standard deviation}
 \item{weight}{population weight}
}
\description{
\code{zscores} calculates standardized scores with a mean close to zero and standard deviation close to one
for the variable given in \code{variable}.
The standardization uses the population weighted mean and standard deviation, which are calculated based on the
population distribution given by \code{population}.
}
\examples{
data <- dep_data

#store all data in object z_oz
z_oz <- zscore(data$total_pop, data$pcnt_overcrowding)

#extract z-score
data$z_overcrowd <- z_oz$z.score
mean(data$z_overcrowd) #mean of z-score
sd(data$z_overcrowd) #sd z-score

#compare weighted mean to actual mean
mean(data$pcnt_overcrowding)
sd(data$pcnt_overcrowding)
}
